+++ src/main/java/com/pnudev/springonlinedelivery/controllers/MenuItemController.java
@@ -1,6 +1,7 @@
 package com.pnudev.springonlinedelivery.controllers;
 
 import com.pnudev.springonlinedelivery.dto.MenuItemDto;
+import com.pnudev.springonlinedelivery.repos.MenuItemRepository;
 import com.pnudev.springonlinedelivery.service.MenuItemService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -18,4 +19,6 @@ public class MenuItemController {
     public List<MenuItemDto> getMenuItems() {
         return menuItemService.getMenuItems();
     }
+
+    private MenuItemRepository menuItemRepository;
 }
