+++ src/main/java/com/pnudev/springonlinedelivery/models/MenuItem.java
@@ -26,5 +26,5 @@ public class MenuItem {
 	
 	@ManyToOne
 	@JoinColumn(name = "order_id", nullable = false)
-	private Order order;	
+	private Order order;
 }
