+++ src/main/java/com/pnudev/springonlinedelivery/service/MenuItemServiceImpl.java
@@ -12,6 +12,7 @@ import java.util.List;
 @AllArgsConstructor
 @Service
 public class MenuItemServiceImpl implements MenuItemService {
+
     private final MenuItemRepository menuItemRepository;
     private final MenuItemMapper menuItemMapper;
 
@@ -37,6 +38,32 @@ public class MenuItemServiceImpl implements MenuItemService {
     @Override
     public void cancelMenuItem(MenuItem menuItem) {
         // TODO TBW
+    }
+
+    public MenuItem dtoToEntity(MenuItemDto menuItemDto) {
+        if (menuItemDto == null) {
+            return null;
+        }
+        MenuItem menuItem = new MenuItem();
+        menuItem.setName(menuItemDto.getName());
+        menuItem.setSummary(menuItemDto.getSummary());
+        menuItem.setSize(menuItemDto.getSize());
+        menuItem.setPrice(menuItemDto.getPrice());
+        menuItem.setOrder(menuItemDto.getOrder());
+        return menuItem;
+    }
+
+    public MenuItemDto entityToDto(MenuItem menuItem) {
+        if (menuItem == null ) {
+            return null;
+        }
+        MenuItemDto menuItemDto = new MenuItemDto();
+        menuItemDto.setName(menuItem.getName());
+        menuItemDto.setSummary(menuItem.getSummary());
+        menuItemDto.setSize(menuItem.getSize());
+        menuItemDto.setPrice(menuItem.getPrice());
+        menuItemDto.setOrder(menuItem.getOrder());
+        return menuItemDto;
     }
     // TODO Add implementation of the list of dishes
 }
